+++ hs/test/HelVM/HelMA/Automata/BrainFuck/FileUtil.hs
@@ -1,0 +1,22 @@
+module HelVM.HelMA.Automata.BrainFuck.FileUtil (
+  readBfFile,
+  readOutFile
+) where
+
+readBfFile :: String -> IO String
+readBfFile fileName = readFile $ buildAbsoluteBfFileName fileName
+
+readOutFile :: String -> IO String
+readOutFile fileName = readFile $ buildAbsoluteOutFileName fileName
+
+buildAbsoluteBfFileName :: String -> String
+buildAbsoluteBfFileName fileName = bfDir <> fileName <> ".bf"
+
+buildAbsoluteOutFileName :: String -> String
+buildAbsoluteOutFileName fileName = bfDir <> "output/" <> fileName <> ".out"
+
+bfDir :: String
+bfDir = dir <> "bf/"
+
+dir :: String
+dir = "examples/"