+++ hs/test/HelVM/HelMA/Automata/SubLeq/FileUtil.hs
@@ -1,0 +1,15 @@
+module HelVM.HelMA.Automata.SubLeq.FileUtil (
+  readSqFile
+) where
+
+readSqFile :: String -> IO String
+readSqFile fileName = readFile $ buildAbsoluteSqFileName fileName
+
+buildAbsoluteSqFileName :: String -> String
+buildAbsoluteSqFileName fileName = sqDir <> fileName <> ".sq"
+
+sqDir :: String
+sqDir = dir <> "sq/"
+
+dir :: String
+dir = "examples/"