+++ docs/CHANGELOG.md
@@ -1,4 +1,14 @@
-# 📅 Revision history for HelCam
+# 📅 Revision history for HelMA
+
+## 0.6.5.0  -- 2021-05-06
+
+* Change name to HelMA
+* Use HSpec
+* Remove HUnit
+* Extract Cat interpreter
+* Add Rev interpreter
+* Add examples
+* Remove gradle
 
 ## 0.6.4.0  -- 2021-03-17